Move stories/ into flame/.storybook. Cleanup useless test
