+++ packages/flame/.storybook/components/SpacedGroup.d.ts
@@ -1,0 +1,4 @@
+import { FlameFlexProps } from '../../src/Core';
+
+export interface SpacedGroupProps extends FlameFlexProps {}
+export const SpacedGroup: React.FC<SpacedGroupProps>;
