+++ packages/flame/.storybook/components/Ul.d.ts
@@ -1,0 +1,1 @@
+export const Ul: React.FC<React.ComponentProps<'ul'>>;
